expose/requirements.txt
boto3
botocore
requests
python-dotenv>=1.0.0
paramiko==3.3.1
pyOpenSSL>=23.2.0
inflect>=7.0.0
